+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/BWGEntities.java
@@ -12,6 +12,7 @@ import net.potionstudios.biomeswevegone.world.entity.oddion.Oddion;
 import net.potionstudios.biomeswevegone.world.entity.pumpkinwarden.PumpkinWarden;
 
 import java.util.function.BiConsumer;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 /**
@@ -36,10 +37,13 @@ public class BWGEntities {
 		return PlatformHandler.PLATFORM_HANDLER.registerEntity(id, factory, category, width, height, trackingRange);
 	}
 
-	public static void registerSpawnPlacements() {
-		SpawnPlacements.register(MAN_O_WAR.get(), SpawnPlacements.Type.IN_WATER, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, ManOWar::checkManOWarSpawnRules);
-		SpawnPlacements.register(ODDION.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.WORLD_SURFACE, Oddion::checkOddionSpawnRules);
+	@SuppressWarnings("unchecked")
+	public static <T extends Mob> void registerSpawnPlacements(Consumer<SpawnPlacement<T>> consumer) {
+		consumer.accept((SpawnPlacement<T>)new SpawnPlacement<>(MAN_O_WAR.get(), SpawnPlacements.Type.IN_WATER, Heightmap.Types.MOTION_BLOCKING_NO_LEAVES, ManOWar::checkManOWarSpawnRules));
+		consumer.accept((SpawnPlacement<T>)new SpawnPlacement<>(ODDION.get(), SpawnPlacements.Type.ON_GROUND, Heightmap.Types.WORLD_SURFACE, Oddion::checkOddionSpawnRules));
 	}
+
+	public record SpawnPlacement<T extends Mob>(EntityType<T> entityType, SpawnPlacements.Type spawnPlacementType, Heightmap.Types heightmapType, SpawnPlacements.SpawnPredicate<T> predicate) {}
 
 	/**
 	 * Registers Entity Attributes