+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/BiomesWeveGoneForge.java
@@ -3,6 +3,7 @@ package net.potionstudios.biomeswevegone.forge;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
+import net.minecraftforge.event.entity.SpawnPlacementRegisterEvent;
 import net.minecraftforge.event.server.ServerAboutToStartEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.DistExecutor;
@@ -20,8 +21,6 @@ import net.potionstudios.biomeswevegone.world.level.levelgen.biome.BWGTerraBlend
 import software.bernie.geckolib.GeckoLib;
 import terrablender.api.SurfaceRuleManager;
 
-import java.util.function.Consumer;
-
 /**
  * Main class for the mod on the Forge platform.
  * @see Mod
@@ -38,7 +37,8 @@ public class BiomesWeveGoneForge {
         MOD_BUS.addListener(this::onInitialize);
         MOD_BUS.addListener(this::onPostInitialize);
         EVENT_BUS.addListener(this::onServerStarting);
-        MOD_BUS.addListener((Consumer<EntityAttributeCreationEvent>) event -> BWGEntities.registerEntityAttributes(event::put));
+        MOD_BUS.addListener((EntityAttributeCreationEvent event) -> BWGEntities.registerEntityAttributes(event::put));
+        MOD_BUS.addListener((SpawnPlacementRegisterEvent event) -> BWGEntities.registerSpawnPlacements(consumer -> event.register(consumer.entityType(), consumer.spawnPlacementType(), consumer.heightmapType(), consumer.predicate(), SpawnPlacementRegisterEvent.Operation.OR)));
         VanillaCompatForge.registerVanillaCompatEvents(EVENT_BUS);
         DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> BiomesWeveGoneClientForge.init(MOD_BUS));
         GeckoLib.initialize();